Cargar datos en la BD, visualizarlos en la tabla, modificar UDFestival.db con las tablas ticket/concierto/cartelera

+++ src/jdo/Cartelera.java
@@ -56,7 +56,7 @@ public class Cartelera {
 	/**
 	 * Obtener el id de la Cartelera
 	 */
-	public int getId() {
+	public long getId() {
 		return id;
 	}
